Allow passing an arrayref as the input of TableReader
Like the shorthand that invokes the Spreadsheet decoder when the
tablereader is given a Workbook object, this allows an arrayref
to automatically invoke the 'Mock' decoder, which then returns the
data as if it had been decoded.

I also renamed the attribute 'data' to 'datasets' to help indicate
it takes an array of tables, not a single table.