Enable per-field coercion
This adds a new attribute Field->coerce. During the validation
step for that column, if the value fails and coerce is enabled,
it runs the coercion and suppresses the validation failure if the
modified value passes the validation.

If the modified value also fails, the original value is preserved
for error messages.