Fix compatibility with Spreadsheet::ParseXLSX 0.34
Prior to 0.34, Spreadsheet::ParseXLSX returned sub-objects in the
Spreadsheet::ParseExcel:: namespace, so a test for one class name
covered both ParseXLSX and ParseExcel.  It was a bad idea to test
specific class names in the first place though, so use duck-typing
to look for the same object methods that detect_input_format looks
for.