Add ->row to decoder iterators, and fix off-by-one errors
The iterators previous has ->position which is supposed to be
free-form text (but reporting 1-based lines) and in many cases it
was returning 0-based numbers and incrementing to a non-existent
line number at EOF.

Now ->row gives a reliable 1-based number, and ->position has been
updated to return text that matches.  For the spreadsheet decoder,
it also now mentions the sheet number for sheets beyond the first.