New API for on_validation_error
on_validation_fail is now on_validation_error so taht I can change
the API of the callback without breaking back-compat.  The new API
gives the user the resulting $record, a $value_ref, the $path to
tat value, and the $data_iterator.

    on_validation_error => sub {
      my ($tablereader, $failures, $record, $data_iterator)= @_;
      # $record is the assembled hashref (unblessed) or arrayref of fields
      # $data_iterator is the Decoder's row iterator, useful for context
      for (@$failures) {
        my ($field, $value_ref, $message, $path)= @$_;
        ...
        # $field is a Data::TableReader::Field
        # $$value_ref is the string that failed validation
        # $message is the error returned from the validation function
        # $path is the element (and maybe sub-element) of $record
        #   i.e.  $value_ref= \$record->{$path[0]}[$path[1]]
        # You may modify $$value_ref or $record to alter the output
      }
      ...
    }

Added unit test for the new API, and verified existing unit tests
for the old API still pass.