Include 'row' in table_search_results->{found}
I previously used row_idx, but that's inconsistent with the
iterators which have 'dataset_idx' and 'row' accessors.
Also, for the case where the headers are implied (and don't
come from any row) it makes more sense for the value to be
undef.

Meanwhile I left row_idx as-is, for back-compat.